kiegészítés a 2dfed16b kommithoz - az adatbázisban is követni kell a típus módosítást, a következő dokumentációk szerint: http://www.postgresql.org/docs/8.4/interactive/datatype-numeric.html http://docs.oracle.com/javase/6/docs/api/constant-values.html#java.lang.Long.MAX_VALUE

+++ schema.sql
@@ -3,7 +3,7 @@
 CREATE SEQUENCE users_seq;
 
 CREATE TABLE users (
-    id              integer PRIMARY KEY DEFAULT nextval('users_seq'),
+    id              bigint PRIMARY KEY DEFAULT nextval('users_seq'),
     usr_nick        varchar(40) NOT NULL,
     usr_name        varchar(60) NOT NULL,
     usr_email       varchar(60) NOT NULL
@@ -12,7 +12,7 @@ CREATE TABLE users (
 CREATE SEQUENCE lessons_seq;
 
 CREATE TABLE lessons (
-    id              integer PRIMARY KEY DEFAULT nextval('lessons_seq'),
+    id              bigint PRIMARY KEY DEFAULT nextval('lessons_seq'),
     ls_name         varchar(60) NOT NULL,
     ls_code         varchar(15) NOT NULL
 );
@@ -20,16 +20,16 @@ CREATE TABLE lessons (
 CREATE SEQUENCE courses_seq;
 
 CREATE TABLE courses (
-    id              integer PRIMARY KEY DEFAULT nextval('courses_seq'),
+    id              bigint PRIMARY KEY DEFAULT nextval('courses_seq'),
     c_code          varchar(15) NOT NULL
 );
 
 CREATE SEQUENCE lessons_courses_seq;
 
 CREATE TABLE lessons_courses (
-    id              integer PRIMARY KEY DEFAULT nextval('lessons_courses_seq'),
-    lesson_id       integer NOT NULL,
-    course_id       integer NOT NULL,
+    id              bigint PRIMARY KEY DEFAULT nextval('lessons_courses_seq'),
+    lesson_id       bigint NOT NULL,
+    course_id       bigint NOT NULL,
     FOREIGN KEY ( lesson_id ) REFERENCES lessons,
     FOREIGN KEY ( course_id ) REFERENCES courses
 );
@@ -37,9 +37,9 @@ CREATE TABLE lessons_courses (
 CREATE SEQUENCE ccRequests_seq;
 
 CREATE TABLE ccRequests (
-    id              integer PRIMARY KEY DEFAULT nextval('ccRequests_seq'),
-    usr_id          integer NOT NULL,
-    course_from_id  integer NOT NULL,
+    id              bigint PRIMARY KEY DEFAULT nextval('ccRequests_seq'),
+    usr_id          bigint NOT NULL,
+    course_from_id  bigint NOT NULL,
     status          varchar(10) NOT NULL,
     FOREIGN KEY ( usr_id ) REFERENCES users,
     FOREIGN KEY ( course_from_id ) REFERENCES courses
@@ -48,9 +48,9 @@ CREATE TABLE ccRequests (
 CREATE SEQUENCE ccRequests_to_courses_seq;
 
 CREATE TABLE ccRequests_to_courses (
-    id              integer PRIMARY KEY DEFAULT nextval('ccRequests_to_courses_seq'),
-    ccReq_id        integer NOT NULL,
-    course_to_id    integer NOT NULL,
+    id              bigint PRIMARY KEY DEFAULT nextval('ccRequests_to_courses_seq'),
+    ccReq_id        bigint NOT NULL,
+    course_to_id    bigint NOT NULL,
     FOREIGN KEY ( ccReq_id ) REFERENCES ccRequests,
     FOREIGN KEY ( course_to_id ) REFERENCES courses
 );
